[ADD] Binary-value on buttonmaps and handle
handle will now print\return the binary representation of the value if
specified in buttonmaps (specified by value being a tuple of
(rawdata-idx, True)).

I realized I needed to add this because the vJoy drivers return 1 value
for buttons no matter how many buttons are pressed at the same time, and
individually they are all exponents of 2, so I figured they were
returned in a binary format. The same with the dpad.